+++ crates/mpz-circuits-generic/src/lib.rs
@@ -1,3 +1,2 @@
-pub mod binary;
 pub mod circuit;
 pub mod model;
